+++ src/main/java/org/openlmis/requisition/service/PermissionService.java
@@ -30,7 +30,7 @@ import java.util.stream.Collectors;
 import org.openlmis.requisition.domain.BaseEntity;
 import org.openlmis.requisition.domain.requisition.Requisition;
 import org.openlmis.requisition.domain.requisition.RequisitionStatus;
-import org.openlmis.requisition.dto.ConvertToOrderDto;
+import org.openlmis.requisition.dto.ReleaseRequisitionLineItemDto;
 import org.openlmis.requisition.dto.ResultDto;
 import org.openlmis.requisition.dto.RightDto;
 import org.openlmis.requisition.dto.UserDto;
@@ -197,16 +197,18 @@ public class PermissionService {
   /**
    * Checks if current user has permission to convert requisition to order.
    *
-   * @param list of ConvertToOrderDtos containing chosen requisitionId and supplyingDepotId.
+   * @param list of ReleaseRequisitionLineItemDto containing requisitionId and supplyingDepotId.
    * @return ValidationResult containing info about the result of this check
    */
-  public ValidationResult canConvertToOrder(List<ConvertToOrderDto> list) {
+  public ValidationResult canConvertToOrder(List<ReleaseRequisitionLineItemDto> list) {
     Map<UUID, Requisition> requisitions = requisitionRepository
-        .findAll(list.stream().map(ConvertToOrderDto::getRequisitionId).collect(Collectors.toSet()))
+        .findAll(list
+            .stream()
+            .map(ReleaseRequisitionLineItemDto::getRequisitionId).collect(Collectors.toSet()))
         .stream()
         .collect(Collectors.toMap(BaseEntity::getId, Function.identity()));
 
-    for (ConvertToOrderDto convertToOrder : list) {
+    for (ReleaseRequisitionLineItemDto convertToOrder : list) {
       Requisition requisition = requisitions.get(convertToOrder.getRequisitionId());
 
       if (requisition == null) {