+++ src/main/java/org/openlmis/requisition/service/RequisitionService.java
@@ -53,7 +53,7 @@ import org.openlmis.requisition.domain.requisition.StatusChange;
 import org.openlmis.requisition.domain.requisition.StatusMessage;
 import org.openlmis.requisition.domain.requisition.StockAdjustmentReason;
 import org.openlmis.requisition.domain.requisition.StockData;
-import org.openlmis.requisition.dto.ConvertToOrderDto;
+import org.openlmis.requisition.dto.ReleaseRequisitionLineItemDto;
 import org.openlmis.requisition.dto.DetailedRoleAssignmentDto;
 import org.openlmis.requisition.dto.FacilityDto;
 import org.openlmis.requisition.dto.IdealStockAmountDto;
@@ -473,8 +473,8 @@ public class RequisitionService {
    *                           order
    * @return list of released requisitions
    */
-  public List<Requisition> releaseRequisitionsAsOrder(
-      List<ConvertToOrderDto> convertToOrderDtos, UserDto user) {
+  private List<Requisition> releaseRequisitionsAsOrder(
+      List<ReleaseRequisitionLineItemDto> convertToOrderDtos, UserDto user) {
     Profiler profiler = new Profiler("RELEASE_REQUISITIONS_AS_ORDER");
     profiler.setLogger(LOGGER);
 
@@ -488,7 +488,7 @@ public class RequisitionService {
         .collect(toSet());
 
     profiler.start("RELEASE");
-    for (ConvertToOrderDto convertToOrderDto : convertToOrderDtos) {
+    for (ReleaseRequisitionLineItemDto convertToOrderDto : convertToOrderDtos) {
       UUID requisitionId = convertToOrderDto.getRequisitionId();
       Requisition loadedRequisition = requisitionRepository.findOne(requisitionId);
       isEligibleForConvertToOrder(loadedRequisition).throwExceptionIfHasErrors();
@@ -507,6 +507,33 @@ public class RequisitionService {
             loadedRequisition.getId()));
       }
 
+      releasedRequisitions.add(loadedRequisition);
+    }
+
+    profiler.stop().log();
+    return releasedRequisitions;
+  }
+
+  /**
+   * Releases the list of given requisitions without creating order.
+   *
+   * @param releaseWithoutOrderDtos list of Requisitions with their supplyingDepots to be released
+   *                           without order.
+   * @return list of released requisitions
+   */
+  private List<Requisition> releaseRequisitionsWithoutOrder(
+      List<ReleaseRequisitionLineItemDto> releaseWithoutOrderDtos) {
+    Profiler profiler = new Profiler("RELEASE_REQUISITIONS_WITHOUT_ORDER");
+    profiler.setLogger(LOGGER);
+
+    List<Requisition> releasedRequisitions = new ArrayList<>();
+
+    profiler.start("RELEASE_WITHOUT_ORDER");
+    for (ReleaseRequisitionLineItemDto convertToOrderDto : releaseWithoutOrderDtos) {
+      UUID requisitionId = convertToOrderDto.getRequisitionId();
+      Requisition loadedRequisition = requisitionRepository.findOne(requisitionId);
+      isEligibleForReleasingWithoutOrder(loadedRequisition).throwExceptionIfHasErrors();
+      loadedRequisition.releaseWithoutOrder(authenticationHelper.getCurrentUser().getId());
       releasedRequisitions.add(loadedRequisition);
     }
 
@@ -563,7 +590,7 @@ public class RequisitionService {
   /**
    * Converting Requisition list to Orders.
    */
-  public void convertToOrder(List<ConvertToOrderDto> list, UserDto user) {
+  public List<Requisition> convertToOrder(List<ReleaseRequisitionLineItemDto> list, UserDto user) {
     Profiler profiler = new Profiler("CONVERT_TO_ORDER");
     profiler.setLogger(LOGGER);
 
@@ -584,7 +611,27 @@ public class RequisitionService {
     orderFulfillmentService.create(orders);
 
     profiler.stop().log();
+    return releasedRequisitions;
   }
+
+  /**
+   * Release requisitions without order.
+   */
+  public List<Requisition> releaseWithoutOrder(List<ReleaseRequisitionLineItemDto> list) {
+    Profiler profiler = new Profiler("RELEASE_WITHOUT_ORDER");
+    profiler.setLogger(LOGGER);
+
+    profiler.start("RELEASE_REQUISITIONS_WITHOUT_ORDER");
+    List<Requisition> releasedRequisitions = releaseRequisitionsWithoutOrder(list);
+
+    for (Requisition requisition : releasedRequisitions) {
+      requisitionRepository.save(requisition);
+      requisitionStatusProcessor.statusChange(requisition);
+    }
+    profiler.stop().log();
+    return releasedRequisitions;
+  }
+
 
   /**
    * Saves status message of a requisition if its draft is not empty.
@@ -669,6 +716,14 @@ public class RequisitionService {
     } else if (!approvedQtyColumnEnabled(requisition)) {
       return ValidationResult.failedValidation(
           ERROR_VALIDATION_CANNOT_CONVERT_WITHOUT_APPROVED_QTY, requisition.getId());
+    }
+    return ValidationResult.success();
+  }
+
+  private ValidationResult isEligibleForReleasingWithoutOrder(Requisition requisition) {
+    if (APPROVED != requisition.getStatus()) {
+      return ValidationResult.failedValidation(
+          ERROR_REQUISITION_MUST_BE_APPROVED, requisition.getId());
     }
     return ValidationResult.success();
   }